Make syntect work as a global plugin
This shouldn't be merged as is, but is a working demo of syntect running
globally; that is, using this plugin a single instance of syntect will
handle highlighting for all open views.

To use this with xi-mac, build-rust-xcode.sh will have to be modified so
that the `--features=global` flag is passed to cargo when compiling
syntect. Alternatively, you can modify syntect/src/main.rs to use the
global code path by default.

Known issues:
- none of the indentation code is ported, so autoindent won't work
- the idle handler has no concept of priority, so if a big file is being
processed in the background text entered in another buffer can end up
not being processed immediately. In some instances it may not be
processed at all.

If eventually merged, this closes #472.